HSEARCH-1673 Moving #getSearchMapping from SearchMappingBuilder to SearchMapping making the builder obsolete.
